backend/requirements.txt
django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
drf-spectacular==0.27.2
Pillow==10.4.0
requests==2.32.3
gunicorn==21.2.0
